requirements.txt
pdfminer.six
Pillow
pytesseract
requests
Unidecode
